### 2021-08-09 #3590 @whedon
[PRE REVIEW]: beamshapes: a Python package to generate directivity patterns for various sound source models
**Submitting author:** @thejasvibr (<a href="http://orcid.org/0000-0001-5360-4383">Beleyur Thejasvi</a>)
**Repository:** <a href="https://github.com/thejasvibr/bat_beamshapes.git" target ="_blank">https://github.com/thejasvibr/bat_beamshapes.git</a>
**Version:** v0.2.0
