### 2021-08-09 #3590 @arfon
If you want to fix the DOIs while you're waiting (noting that Whedon's suggestions are not always right), you can, then use the following commands (one at a time, as the first line of a new comment) to regenerate the PDF and check the references.
