### 2021-08-11 0cf727f
add details to CONTRIBUTING, and generate a commit to mark removal of copyrighted material from history
