### 2021-09-06 #3590 @faroit
Hi @thejasvibr, sorry for the long waiting time. I wanted to make sure that I have enough bandwidth to probably maintain the editing process before kicking this off.

> Thanks for agreeing to edit @faroit. Suggested reviewers from my side are DrMarc, pharshalp, astanziola

I think that sounds good, I would also like to include an audio researcher who isn't that much familiar with animal sounds but general dsp and python: I would `nils-werner` if there is not conflict of interest.

In the meantime, I suggest you can already work on the missing DOI as pointed out by @arfon
