### 2021-09-08 #3590 @faroit
👋  @nils-werner @DrMarc - would any of you be willing to review this submission for JOSS? The JOSS review process takes place on GitHub and focuses on the software and a short paper. We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

The software under review is [beamshapes: a Python package to generate directivity patterns for various sound source models](https://github.com/thejasvibr/bat_beamshapes.git).
