### 2021-09-19 8e1b93c
update paper.bib [skip travis]
include DOIs as raised by whedon.
