+++ paper.bib
@@ -2,14 +2,16 @@
   title={Acoustics: sound fields and transducers},
   author={Beranek, Leo Leroy and Mellow, Tim},
   year={2012},
-  publisher={Academic Press}
+  publisher={Academic Press},
+  doi={10.1016/c2011-0-05897-0}
 }
 
 @book{beranek2019acoustics,
   title={Acoustics: Sound Fields, Transducers and Vibration},
   author={Beranek, Leo and Mellow, Tim},
   year={2019},
-  publisher={Academic Press}
+  publisher={Academic Press},
+  doi={10.1016/c2017-0-01630-0}
 }
 
 
@@ -21,7 +23,8 @@
   number={6B},
   pages={1430--1432},
   year={1970},
-  publisher={Acoustical Society of America}
+  publisher={Acoustical Society of America},
+  doi={10.1121/1.1912304}
 }
 
 
@@ -33,7 +36,8 @@
   number={2},
   pages={165--171},
   year={1979},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/bf00610475}  
 }
 
 
@@ -56,7 +60,8 @@
   number={6},
   pages={4175--4188},
   year={2020},
-  publisher={Acoustical Society of America}
+  publisher={Acoustical Society of America},
+  doi={10.1121/10.0001376}
 }
 
 @article{jakobsen2013convergent,
@@ -67,7 +72,8 @@
   number={7430},
   pages={93--96},
   year={2013},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+   doi={10.1038/nature11664}
 }
 
 @article{kounitsky2015bats,
@@ -78,7 +84,8 @@
   number={21},
   pages={6724--6729},
   year={2015},
-  publisher={National Acad Sciences}
+  publisher={National Acad Sciences},
+  doi={10.1073/pnas.1422843112}  
 }
 
 
@@ -90,7 +97,8 @@
   number={2},
   pages={1046--1058},
   year={2011},
-  publisher={Acoustical Society of America}
+  publisher={Acoustical Society of America},
+  doi={10.1121/1.3519408}
 }
 
 
@@ -172,7 +180,8 @@
   volume={170},
   pages={103996},
   year={2020},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.beproc.2019.103996}
 }
 
 @article{gerhardt1975sound,
@@ -183,7 +192,8 @@
   number={1},
   pages={1--12},
   year={1975},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/bf00657481}
 }
 
 @article{patricelli2008acoustic,
@@ -194,7 +204,8 @@
   number={4},
   pages={1389--1401},
   year={2008},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.anbehav.2008.07.005}
 }
 
 @article{witkin1977importance,
@@ -205,7 +216,8 @@
   number={4},
   pages={490--493},
   year={1977},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi={10.2307/1367730}
 }
 
 @article{yorzinski2010birds,
@@ -217,6 +229,7 @@
   pages={923--932},
   year={2010},
   publisher={The Royal Society}
+  doi={10.1098/rspb.2009.1519}
 }
 
 @article{larsen1990directionality,
@@ -225,7 +238,8 @@
   journal={Ornis Scandinavica},
   pages={37--45},
   year={1990},
-  publisher={JSTOR}
+  publisher={JSTOR},
+  doi={10.2307/3676376}
 }
 
[…]
