+++ paper.bib
@@ -228,7 +228,7 @@
   number={1683},
   pages={923--932},
   year={2010},
-  publisher={The Royal Society}
+  publisher={The Royal Society},
   doi={10.1098/rspb.2009.1519}
 }
 
