+++ paper.bib
@@ -49,7 +49,8 @@
   number={1658},
   pages={853--860},
   year={2009},
-  publisher={The Royal Society London}
+  publisher={The Royal Society London},
+  doi={10.1098/rspb.2008.1505 }
 }
 
 @article{macaulay2020high,
@@ -154,7 +155,9 @@
   volume={3},
   pages={e103},
   year={2017},
-  publisher={PeerJ Inc.}
+  publisher={PeerJ Inc.},
+  url = {https://doi.org/10.7717/peerj-cs.103},
+ doi = {10.7717/peerj-cs.103}
 }
 
 @manual{mpmath,
@@ -166,11 +169,11 @@
   year    = {2021},
 }
 
-@article{hart2011flint,
+@manual{hart2011flint,
   title={FLINT--Fast Library for Number Theory},
   author={Hart, William and Johansson, Fredrik and Pancratz, Sebastian},
   year={2011},
-  publisher={Citeseer}
+  publisher={Citeseer},
 }
 
 @article{rodriguez2020sound,
