### 2021-09-20 #3740 @whedon
[REVIEW]: beamshapes: a Python package to generate directivity patterns for various sound source models
**Submitting author:** @thejasvibr (<a href="http://orcid.org/0000-0001-5360-4383">Beleyur Thejasvi</a>)
**Repository:** <a href="https://github.com/thejasvibr/bat_beamshapes" target ="_blank">https://github.com/thejasvibr/bat_beamshapes</a>
**Version:** v0.2.2
