### 2021-09-21 thejasvibr/bat_beamshapes#2 @nils-werner
Package is not pip-installable
Hi, this issue is related to your submission to JOSS, [#3740](https://github.com/openjournals/joss-reviews/issues/3740).

I have noticed that your package is not easily pip-installable, i.e.

```
git clone https://github.com/thejasvibr/bat_beamshapes.git
pip install -e bat_beamshapes
```

fails with

```
     command: /Users/nwerner/Projekte/beamshapes/bin/python3.7 -c 'import sys, setuptools, tokenize; sys.argv[0] = '"'"'/Users/nwerner/Projekte/beamshapes/bat_beamshapes/setup.py'"'"'; __file__='"'"'/Users/nwerner/Projekte/beamshapes/bat_beamshapes/setup.py'"'"';f=getattr(tokenize, '"'"'open'"'"', open)(__file__);code=f.read().replace('"'"'\r\n'"'"', '"'"'\n'"'"');f.close();exec(compile(code, __file__, '"'"'exec'"'"'))' egg_info --egg-base /private/var/folders/gf/gn2yxq511p77gzs2gsn4gp2w0000gn/T/pip-pip-egg-info-b4i9yk__
         cwd: /Users/nwerner/Projekte/beamshapes/bat_beamshapes/
    Complete output (9 lines):
    Traceback (most recent call last):
      File "<string>", line 1, in <module>
      File "/Users/nwerner/Projekte/beamshapes/bat_beamshapes/setup.py", line 3, in <module>
        import beamshapes
      File "/Users/nwerner/Projekte/beamshapes/bat_beamshapes/beamshapes/__init__.py", line 8, in <module>
        from beamshapes.piston_in_sphere import piston_in_sphere_directivity
      File "/Users/nwerner/Projekte/beamshapes/bat_beamshapes/beamshapes/piston_in_sphere.py", line 34, in <module>
        from joblib import Parallel, delayed
    ModuleNotFoundError: No module named 'joblib'
ERROR: Command errored out with exit status 1: python setup.py egg_info Check the logs for full command output.
```

I see that you have posted different installation instructions using a `requirements.txt` file, but since you are planning to release the package to PyPI, plain pip-installability will eventually become a requirement. Therefore I'd recommend you to fix this.
