### 2021-09-21 thejasvibr/bat_beamshapes#3 @nils-werner
Code cleanliness
Hi, this issue is related to your submission to JOSS, [#3740](https://github.com/openjournals/joss-reviews/issues/3740).

I have noticed that your codebase has a lot of commented out code blocks:

 - [`__init__.py`](https://github.com/thejasvibr/bat_beamshapes/blob/dev/beamshapes/__init__.py#L6-L7)
 - [`sim_reclevels.py`](https://github.com/thejasvibr/bat_beamshapes/blob/dev/beamshapes/sim_reclevels.py#L155-L201)
 - [`piston_in_sphere_flint.py`](https://github.com/thejasvibr/bat_beamshapes/blob/dev/beamshapes/piston_in_sphere_flint.py#L300-L332)
 - and many more.

Also, many of your files have Jupyter/VS Code notebook cell makers, like

```
# %%
```

I understand that in many situations, library code evolves from experimental notebook-code, but I would still like to suggest to give the codebase a proper cleanup before publication.
