### 2021-09-21 thejasvibr/bat_beamshapes#4 @nils-werner
Link to documentation broken
Hi, this issue is related to your submission to JOSS, [#3740](https://github.com/openjournals/joss-reviews/issues/3740).

Currently your `README` links to the docs under

```
https://github.com/thejasvibr/bat_beamshapes/blob/dev/beamshapes.rtfd.io
```

while I believe the correct address should be

```
https://beamshapes.rtfd.io
```

Also to improve visibility of the docs I would also suggest to add this URL to the Repo info, and possibly also add an [RDT badge](https://docs.readthedocs.io/en/stable/badges.html)
