### 2021-09-21 thejasvibr/bat_beamshapes#5 @nils-werner
API docs output broken?
Hi, this issue is related to your submission to JOSS, [#3740](https://github.com/openjournals/joss-reviews/issues/3740).

In some places the API docs appear to have broken output: e.g. on https://beamshapes.readthedocs.io/en/latest/source_models.html#beamshapes.point_source_on_a_sphere.d_theta_func it looks like this:

![Screenshot 2021-09-21 at 13 26 54](https://user-images.githubusercontent.com/88704/134162652-af00eed2-588a-4966-8011-1e814ce090aa.png)

Is this on purpose?
