### 2021-09-21 thejasvibr/bat_beamshapes#6 @nils-werner
Cannot run code examples
Hi, this issue is related to your submission to JOSS, [#3740](https://github.com/openjournals/joss-reviews/issues/3740).

I tried running the code example on https://beamshapes.readthedocs.io/en/latest/gallery_examples/0_preprint_figures.html

and it fails with

```
Traceback (most recent call last):
  File "scribble.py", line 5, in <module>
    from beamshapes import piston_in_sphere_flint as piston_sphere # to speed things up!!
  File "/Users/nwerner/Projekte/beamshapes/bat_beamshapes/beamshapes/piston_in_sphere_flint.py", line 28, in <module>
    import flint
ModuleNotFoundError: No module named 'flint'
```

I would recommend to add and use [doctest](https://docs.python.org/3/library/doctest.html)s or [pytest doctest](https://docs.pytest.org/en/latest/how-to/doctest.html)s to ensure code examples are runnable and reproducible.
