### 2021-09-21 thejasvibr/bat_beamshapes#7 @nils-werner
Docs are very terse
Hi, this issue is related to your submission to JOSS, [#3740](https://github.com/openjournals/joss-reviews/issues/3740).

I have noticed the docs for your library are very terse. While you do have a section that roughly explains the concept of a directivity pattern, you very quickly dive into technicalities like parameters and specific patterns. This makes it really hard for people inside and outside of your field to quickly understand the purpose of this library and what problems it solves.

I would recommend expanding on these sections, using more text and simpler language.

Also, the implemented directivity patterns are [only listed by name and a single figure each](https://beamshapes.readthedocs.io/en/latest/general_intro.html#source-models). Just from these figures, I was unfortunately not able to understand the actual shape and its properties. Again, I would recommend expanding on these sections, and maybe more figures.
