### 2021-09-21 thejasvibr/bat_beamshapes#8 @nils-werner
Unittests are very dense
Hi, this issue is related to your submission to JOSS, [#3740](https://github.com/openjournals/joss-reviews/issues/3740).

This is more a personal remark and wouldn't be a reason for rejection.

I noticed that your unit test code is very dense and contains a lot of boilerplate, which makes it really hard to understand the tests. I believe a useful unittest should not only successfully test a feature, but also be simple enough so that developers can quickly understand what feature is being tested (i.e. to recognize holes in test coverage).

To that end I would recommend cleaning up and slimming down the test code, and potentially moving to a modern test framework, like [pytest](https://docs.pytest.org/).
