### 2021-09-22 thejasvibr/bat_beamshapes#2 @thejasvibr
Thanks a lot @nils-werner  for the raised issues, and helpful feedback. Will get back to resolving and replying to each over the coming week or two!
