### 2021-10-04 thejasvibr/bat_beamshapes#2 @thejasvibr
pip installation doesn't work:

```
from joblib import Parallel, delayed
    ModuleNotFoundError: No module named 'joblib'
```
possible error in ```import beamshapes``` in setup.py.
The __init__.py hase the ```__version__``` number, but also imports other modules that require dependencies.
