### 2021-10-05 8d5721f
add version.py file to have a single package version number
