+++ beamshapes/__init__.py
@@ -2,7 +2,8 @@
 All others will need to be deliberately loaded. 
 '''
 
-__version__ = '0.2.0'
+from beamshapes.version import __version__
+
 # from beamshapes.beamshape_predictions import * 
 # from beamshapes.sim_reclevels  import * 
 from beamshapes.piston_in_sphere import piston_in_sphere_directivity
