### 2021-10-05 8f442d5
up the version number by one patch because test pypi isn't recognising the change
