+++ beamshapes/version.py
@@ -1,1 +1,1 @@
-__version__ = '0.2.0a'
+__version__ = '0.2.1'
