### 2021-10-05 853965c
upload onto pypi and alter README to reflect that
