+++ setup.py
@@ -20,9 +20,9 @@ setup(name='beamshapes',
      author='Thejasvi Beleyur',
      author_email='thejasvib@gmail.com',
      license='MIT',
-     packages=find_packages(),
      install_requires=['joblib','numpy','sympy','mpmath',
         'scipy','matplotlib','tqdm'],
+          packages=find_packages(),
      zip_safe=False,
      include_package_data=True,
      classifiers=[
