### 2021-10-05 thejasvibr/bat_beamshapes#2 @thejasvibr
The package is now locally installable, and also on PyPi for general installation. Have tested it on two separate Windows systems, will try it out on a Linux at some point - hope it will work there too.
