### 2021-10-05 thejasvibr/bat_beamshapes#4 @thejasvibr
Thanks for pointing it out. Have updated the broken link, and also included the RTD docs badge to signal latest status.
