### 2021-10-06 thejasvibr/bat_beamshapes#5 @thejasvibr
Hi,
It does look messy indeed! This isn't really on purpose, but rather the side effect of having SymPy-derived lambda-s using [lambdify](https://docs.sympy.org/latest/modules/utilities/lambdify.html). SymPy-derived lambdas convert math expressions into functions that accept concrete numerical inputs. The messy looking output is the underlying code implementation of the math expression that is returned on doing eg. ```help(d_theta_func)```.

The whole package has a lot of such Sympy-derived lambdas - do you think it's better to get rid of them for visual cleanliness?
