### 2021-10-07 7ab6f7d
clean up tests and add testing.md
