+++ beamshapes/tests/test_osccapsphere.py
@@ -21,36 +21,6 @@ import tqdm
 from beamshapes import cap_in_sphere_directivity
 from beamshapes.cap_in_sphere import d_zero
 
-# class CapOfSphere(unittest.TestCase):
-    
-#     def setUp(self):
-#         self.paramv = {}
-#         plotdata = pd.read_csv('plots_data/cap-sphere_fig12-27.csv')        
-#         self.by_ka = plotdata.groupby('ka')
-    
-#     def perform_ka_match(self, kaval):
-#         self.paramv['k'] = 10
-#         self.paramv['R'] = 0.01
-#         self.paramv['alpha'] = mpmath.pi/3
-#         self.paramv['a'] = self.paramv['R']*mpmath.sin(self.paramv['alpha'])
-#         self.paramv['k'] = kaval/self.paramv['a']
-#         self.paramv['a'] = kaval/self.paramv['k']
-#         angles = np.radians(self.by_ka.get_group(kaval)['theta'])
-#         actual_dirnlty = self.by_ka.get_group(kaval)['r'].to_numpy()
-        
-#         _, output_dirnlty = cap_in_sphere_directivity(angles,
-#                                                                       self.paramv)
-#         error = np.abs(output_dirnlty-actual_dirnlty)
-#         return np.max(error)
-    
-#     def test_kamatches(self):
-#         kavals = [1, 3, 10, 30]
-#         max_abs_errors = np.zeros(len(kavals))
-#         for i, each in enumerate(kavals):
-#             max_abs_errors[i] = self.perform_ka_match(each)
-#         print(max_abs_errors)
-#         self.assertTrue(np.all(max_abs_errors<=1.05))
-
 
 class OnAxisCapOfSphere(unittest.TestCase):
     '''
@@ -58,12 +28,14 @@ class OnAxisCapOfSphere(unittest.TestCase):
     '''
     
     def setUp(self):
+        # load the ground-truth data and parameters used for the results in the plots
         self.paramv = {}
         self.plotdata = pd.read_csv('plots_data/onaxis_capofsphere_v2.csv')        
         self.alpha_values = self.plotdata['alpha_deg'].unique()
         self.by_alpha = self.plotdata.groupby('alpha_deg')
         
     def test_onaxis_match(self):
+        # set fixed radius
         self.paramv['R'] = 0.1
         groundtruth_levels = self.plotdata['onaxis_db'].tolist()
         all_obtained_raw = []
@@ -80,7 +52,6 @@ class OnAxisCapOfSphere(unittest.TestCase):
                 this_alpha_set.append(float(np.abs(onaxis_value)))
 
             this_alpha_normalised = np.array(this_alpha_set)
-            #this_alpha_normalised *= 1/np.min(this_alpha_normalised)
             all_obtained_raw.append(this_alpha_normalised)
         
         all_obtained_raw =  np.concatenate(all_obtained_raw)
@@ -91,53 +62,3 @@ class OnAxisCapOfSphere(unittest.TestCase):
 
 if __name__=='__main__':
     unittest.main()
-
-    # #%% 
-    # import matplotlib.pyplot as plt
-    
-    # paramv = {}
-    # paramv['R'] = 0.1
-    # plotdata = pd.read_csv('plots_data/onaxis_capofsphere_v2.csv')
-    # groundtruth_levels = plotdata['onaxis_db']
-    # all_obtained = []
-    # all_obtained_raw = {}
-    # alpha_values = plotdata['alpha_deg'].unique()
-    # by_alpha = plotdata.groupby('alpha_deg')
-    # all_kas = {}
-    # for alpha in alpha_values:
[…]
