+++ beamshapes/tests/test_pointsourcesphere.py
@@ -69,7 +69,3 @@ class OnAxisPointSourceOnSphere(unittest.TestCase):
 
 if __name__=='__main__':
     unittest.main()        
-    
-    
-    
-        
