+++ beamshapes/tests/test_pistoninsphereflint.py
@@ -83,49 +83,3 @@ except:
 
 if __name__=='__main__':
     unittest.main()
-    
-    # #%%
-    # plotdata = pd.read_csv('plots_data/pistoninsphere.csv')        
-    # by_ka = plotdata.groupby('ka')
-    # #angles = np.radians(by_ka.get_group(float(kaval))['theta_deg'])
-    # #actual_dirnlty = by_ka.get_group(float(kaval))['relonaxis_db'].to_numpy()
-    
-    # #%% 
-    # paramv = {}
-    # plotdata = pd.read_csv('plots_data/pistoninsphere.csv')        
-    # by_ka = plotdata.groupby('ka')
-    # frequency = acb(50*10**3) # kHz
-    # vsound = acb(330) # m/s
-    # wavelength = vsound/frequency
-    # alpha_value = pi/3 # 60 degrees --> pi/3
-    # k_value = 2*pi/(wavelength)
-    # paramv = {}
-    # paramv['alpha'] = alpha_value
-    # paramv['k'] = k_value
-    # #%%
-    # #print(f'Starting piston in sphere for ka={kaval}')
-    # kaval = 10
-    # ka = acb(kaval)
-    # a_value = ka/paramv['k']
-    # R_value = a_value/sin(paramv['alpha'])  # m
-    # paramv['R'] = R_value        
-    # paramv['a'] = a_value
-    # paramv['a'] = ka/paramv['k']
-    # angles = np.radians(by_ka.get_group(kaval)['theta_deg'])
-    # angles_acb = [acb(each) for each in angles]
-    # actual_dirnlty = by_ka.get_group(kaval)['relonaxis_db'].to_numpy()
-    
-    # _, output_dirnlty = piston_in_sphere_directivity(angles_acb,
-    #                                                   paramv)
-    # error = np.abs(output_dirnlty-actual_dirnlty)
-    
-    # #%%
-    # import matplotlib.pyplot as plt
-    # plt.figure()
-    # a0 = plt.subplot(111, projection='polar')
-    # plt.plot(angles, output_dirnlty, label='output')
-    # plt.plot(angles, actual_dirnlty, label='ground-truth')
-    # #plt.ylim(-60,0);plt.yticks(np.arange(-60,10,10))
-    # #%%
-    # plt.figure()
-    # plt.plot(np.abs(output_dirnlty-actual_dirnlty))

[…]
