### 2021-10-15 thejasvibr/bat_beamshapes#9 @hagenw
No documentation on how to run tests
This is related to https://github.com/openjournals/joss-reviews/issues/3740

I would recommend to add a section to `CONTRIBUTING` or `README` on how to execute the unit tests.

I also found `beamshapes/tests/testing.md`, which should maybe also linked to at some place.

I did execute the following:

```bash
$ python -m unittest
```

Hope this is the desired way to run the tests?
