### 2021-10-15 thejasvibr/bat_beamshapes#11 @hagenw
DOC: table of contents looks broken
Related to https://github.com/openjournals/joss-reviews/issues/3740

The gap between the caption and the actual sections is too big and it is very hard to relate the sections to the right cpations:

![image](https://user-images.githubusercontent.com/173624/137478848-5d582906-5db8-4ea4-84a8-9a0546476bd4.png)
