### 2021-10-15 thejasvibr/bat_beamshapes#12 @hagenw
Unit tests should indicate the code coverage
Related to https://github.com/openjournals/joss-reviews/issues/3740

Test packages like `pytest` usually provide the possibility to measure the code coverage of the tests. A code coverage of 100% does not mean the code is really well enough tested, but at least it is an indication that the developers have worked on the tests.

At the moment I don't see an easy presentation of the code coverage for the unit tests performed in this package. Would be nice if this could be added at least to the output in the command line when running the tests.
