### 2021-10-15 thejasvibr/bat_beamshapes#13 @hagenw
DOC: examples do not work out of the box
Related to https://github.com/openjournals/joss-reviews/issues/3740

When going to the first example page at https://beamshapes.readthedocs.io/en/latest/gallery_examples/0_preprint_figures.html#sphx-glr-gallery-examples-0-preprint-figures-py

and then trying to run the presented code it does fail already during import with:

```
ModuleNotFoundError: No module named 'flint'
```

I would recommend to state which additional steps the user has to do in order to execute the examples.
