### 2021-10-15 thejasvibr/bat_beamshapes#14 @hagenw
Add instructions to CONTRIBUTING on how to build the docs
Related to https://github.com/openjournals/joss-reviews/issues/3740

There is also the file `docs/README.md`, but that contains information not related to the documentation folder.

I tired with:

```bash
$ pip install -r docs/docs_requirements.txt
```
and then
```bash
$ python -m sphinx docs/ build/html -b html
```
which is the usual way to build documentation with `sphinx`, but this fails with
```
Application error:
config directory doesn't contain a conf.py file (docs/)
```

What I ended up doing in the end is:

```bash
$ cd docs
$ make html
```
but I'm not sure if this is the correct way to build the docs.
