### 2021-10-15 thejasvibr/bat_beamshapes#13 @hagenw
Duplicate of #6
