### 2021-10-15 thejasvibr/bat_beamshapes#15 @hagenw
Cannot build documentation
Related to https://github.com/openjournals/joss-reviews/issues/3740

I'm not sure if I tried it the right way (see #14), but when building the documentation with:

```bash
$  pip install -r docs/docs_requirements.txt
$ cd docs
$ make html
```
I get the following errors and warnings

```
Running Sphinx v3.5.3
making output directory... done
generating gallery...
Using Sphinx-Gallery to convert rst text blocks to markdown for .ipynb files.
generating gallery for gallery_examples... [100%] pistoon_inf_baffle.py
computation time summary:
    - ../../examples/0_preprint_figures.py:    0.00 sec   0.0 MB
    - ../../examples/1_piston_sphere_eg1.py:   0.00 sec   0.0 MB
    - ../../examples/cap_sphere.py:            0.00 sec   0.0 MB
    - ../../examples/pistoon_inf_baffle.py:    0.00 sec   0.0 MB
building [mo]: targets for 0 po files that are out of date
building [html]: targets for 9 source files that are out of date
updating environment: [new config] 14 added, 0 changed, 0 removed
reading sources... [100%] source_models
/home/audeering.local/hwierstorf/git/bat_beamshapes/docs/source/index.rst:86: WARNING: toctree contains reference to nonexisting document ':maxdepth:1'
WARNING: autodoc: failed to import module 'flint_parallelisation' from module 'beamshapes'; the following exception was raised:
Traceback (most recent call last):
  File "/home/audeering.local/hwierstorf/.envs/bat_beamshapes/lib/python3.6/site-packages/sphinx/ext/autodoc/importer.py", line 71, in import_module
    return importlib.import_module(modname)
  File "/usr/lib/python3.6/importlib/__init__.py", line 126, in import_module
  File "<frozen importlib._bootstrap>", line 994, in _gcd_import
  File "<frozen importlib._bootstrap>", line 971, in _find_and_load
  File "<frozen importlib._bootstrap>", line 955, in _find_and_load_unlocked
  File "<frozen importlib._bootstrap>", line 665, in _load_unlocked
  File "<frozen importlib._bootstrap_external>", line 678, in exec_module
  File "<frozen importlib._bootstrap>", line 219, in _call_with_frames_removed
  File "/home/audeering.local/hwierstorf/.envs/bat_beamshapes/lib/python3.6/site-packages/beamshapes-0.2.1-py3.6.egg/beamshapes/flint_parallelisation.py", line 15, in <module>
    acb = flint.acb
AttributeError: module 'flint' has no attribute 'acb'

/home/audeering.local/hwierstorf/git/bat_beamshapes/docs/source/notes_piston_on_prolate_spheroid.rst:19: WARNING: Bullet list ends without a blank line; unexpected unindent.
/home/audeering.local/hwierstorf/git/bat_beamshapes/docs/source/notes_piston_on_prolate_spheroid.rst:39: WARNING: Bullet list ends without a blank line; unexpected unindent.
/home/audeering.local/hwierstorf/git/bat_beamshapes/docs/source/notes_rectcap_of_sphere.rst:149: WARNING: Title underline too short.

Computational implementation of |I0n|:
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
/home/audeering.local/hwierstorf/git/bat_beamshapes/docs/source/notes_rectcap_of_sphere.rst:149: WARNING: Title underline too short.

Computational implementation of |I0n|:
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
WARNING: autodoc: failed to import module 'piston_in_sphere_flint' from module 'beamshapes'; the following exception was raised:
Traceback (most recent call last):
  File "/home/audeering.local/hwierstorf/.envs/bat_beamshapes/lib/python3.6/site-packages/sphinx/ext/autodoc/importer.py", line 71, in import_module
    return importlib.import_module(modname)
  File "/usr/lib/python3.6/importlib/__init__.py", line 126, in import_module
  File "<frozen importlib._bootstrap>", line 994, in _gcd_import
  File "<frozen importlib._bootstrap>", line 971, in _find_and_load
  File "<frozen importlib._bootstrap>", line 955, in _find_and_load_unlocked
  File "<frozen importlib._bootstrap>", line 665, in _load_unlocked
  File "<frozen importlib._bootstrap_external>", line 678, in exec_module
  File "<frozen importlib._bootstrap>", line 219, in _call_with_frames_removed
  File "/home/audeering.local/hwierstorf/.envs/bat_beamshapes/lib/python3.6/site-packages/beamshapes-0.2.1-py3.6.egg/beamshapes/piston_in_sphere_flint.py", line 32, in <module>
    from beamshapes.flint_parallelisation import conv_acb_to_str, conv_str_acb, interchange_params_and_str
  File "/home/audeering.local/hwierstorf/.envs/bat_beamshapes/lib/python3.6/site-packages/beamshapes-0.2.1-py3.6.egg/beamshapes/flint_parallelisation.py", line 15, in <module>
    acb = flint.acb
AttributeError: module 'flint' has no attribute 'acb'

looking for now-outdated files... none found
pickling environment... done
checking consistency... /home/audeering.local/hwierstorf/git/bat_beamshapes/docs/source/general_workflow.rst: WARNING: document isn't included in any toctree
/home/audeering.local/hwierstorf/git/bat_beamshapes/docs/source/notes_piston_in_sphere.rst: WARNING: document isn't included in any toctree
/home/audeering.local/hwierstorf/git/bat_beamshapes/docs/source/notes_piston_on_prolate_spheroid.rst: WARNING: document isn't included in any toctree
/home/audeering.local/hwierstorf/git/bat_beamshapes/docs/source/notes_rectcap_of_sphere.rst: WARNING: document isn't included in any toctree
done
preparing documents... done
writing output... [100%] source_models
generating indices... genindex py-modindex done
highlighting module code... [100%] beamshapes.utilities
writing additional pages... search done
copying images... [100%] _static/pistoninsphere_deviation_2021-05-30_ka=3.png
copying downloadable files... [100%] gallery_examples/pistoon_inf_baffle.ipynb
copying static files... done
copying extra files... done
dumping search index in English (code: en)... done
dumping object inventory... done
build succeeded, 11 warnings.

The HTML pages are in build/html.

Sphinx-Gallery successfully executed 0 out of 0 files subselected by:

    gallery_conf["filename_pattern"] = '\\/plot'
    gallery_conf["ignore_pattern"]   = '__init__\\.py'

after excluding 0 files that had previously been run (based on MD5).

embedding documentation hyperlinks...
embedding documentation hyperlinks for gallery_examples... [100%] cap_sphere.html
```
