### 2021-10-15 thejasvibr/bat_beamshapes#5 @hagenw
I also thought your math output is simply broken in the API docs.

I would recommend to find a way to render the math output as math.

Also the there are lots of `:` without content afterwards and one wonders if the  `_lambdifygenerated()` is part of the API or something used to generate the documentation.

So in general I would recommend to revisit the API documentation.
