### 2021-10-15 thejasvibr/bat_beamshapes#7 @hagenw
I agree and would also recommend to improve the introduction of the package.
