### 2021-10-15 thejasvibr/bat_beamshapes#19 @hagenw
Relation to other more room acoustic relates Python packages
Related to https://github.com/openjournals/joss-reviews/issues/3740

In the paper you state

> To my knowledge there are no computational implementations of sound
> source directivities that are open-source and developed using modern software practices such
> as version-control and unit-testing

which seems to be correct to my knowledge.

But to foster more interaction with the non-bioacoustic community it would be great to add a few statements how it relates to other Python packages, that include some kind of directivity estimations, e.g. https://pyroomacoustics.readthedocs.io/en/pypi-release/pyroomacoustics.directivities.html or https://github.com/AppliedAcousticsChalmers/levitate (see https://github.com/AppliedAcousticsChalmers/levitate/blob/c6ad1038327abfc82a5402b42019b69e52f7f5c4/levitate/transducers.py#L720).
