### 2021-10-18 thejasvibr/bat_beamshapes#9 @thejasvibr
Hi @hagenw ,thanks for the detailed issues. Much to clarify and fix clearly.
Will get down to working on them over the course of the next month!
