### 2021-10-19 bb751dc
closes #9; Include running and writing unit-tests in CONTRIBUTING.md
