### 2021-10-19 thejasvibr/bat_beamshapes#9 @thejasvibr
Have included instructions to run and write tests with links as of this commit.
