### 2021-10-19 thejasvibr/bat_beamshapes#14 @thejasvibr
Instructions included on building the docs with detailed steps in the ```CONTRIBUTING.md```.
