### 2021-10-19 44ac3c5
closes #14; travis skip; included explicit instructions on generating local docs
