+++ CONTRIBUTING.md
@@ -20,9 +20,24 @@ For reasons of code homogeneity, I request that you keep the template structure 
 
 ##  Unit-tests
 ### Running unit-tests
-To run the unittests, first clone an install the ````beamshapes``` package (see the [README](README.md)). From the root of the directory, open your Python command line tool of choice and run:
+To run the unittests, first clone and install the ````beamshapes``` package (see the [README](README.md)).To run the tests you will need to install all the packages in the ```tests/requirements_test.txt``` file. From the root of the directory, open your Python command line tool of choice and run:
 
-```python -m unittest``` to run unit tests. 
+```python -m unittest```  
 
 ### Writing unit-tests
 The unit-tests in ```beamshapes``` should test the computational implementation for correctness based on previously published results. See [here](beamshapes/tests/testing.md) for more on the testing approach. 
+
+## Documentation
+
+## Building docs locally
+To build the documentation locally, you must first install all the required packages:
+
+```pip install -r docs/docs_requirements.txt``` 
+
+And then:
+
+* Move into the ```docs``` folder with ```cd docs```
+* Make an empty ```build``` folder with ```mkdir build```
+* Generate the docs with ```sphinx-build source build```
+
+All the formatted web-pages are in the ```build/``` directory. Start with the ```index.html``` to begin browsing from the landing page. 
