### 2021-10-19 eb6b402
closes #18; skip travis
