### 2021-10-19 thejasvibr/bat_beamshapes#11 @thejasvibr
This seems to be an active issue with the sphinx_rtd_theme  ([issue link](https://github.com/readthedocs/sphinx_rtd_theme/issues/1029)). Not sure what I can do to fix it right now - will search around for some time to see if there are any fixes.
