### 2021-10-19 thejasvibr/bat_beamshapes#12 @thejasvibr
Did not find a direct way to estimate code coverage in the unittest package. Did find the [coverage](https://coverage.readthedocs.io/en/6.0.2/index.html) package, and have included steps to estimate code coverage in CONTRIBUTING.md.

Hope this resolves this issue?
