### 2021-10-19 aba130c
skip travis; give details on test coverage
