+++ docs/source/conf.py
@@ -64,6 +64,7 @@ html_theme_options = {
     # Toc options
     'collapse_navigation': False,
     'sticky_navigation': True,
+    'navigation_depth':5
 }
 
 
