### 2021-10-21 a88ac17
add ref to pyroomac. and levitate; travis skip; closes #19
