+++ paper.md
@@ -41,7 +41,7 @@ Directivity functions can be used in a two-fold manner 1) to deliberately engine
 
 # Statement of need
 
-A host of published directivity functions exist in the literature, but it is my experience that their computational implementations remain as inhouse scripts that are often in proprietary language platforms. To my knowledge there are no computational implementations of sound source directivities that are open-source and developed using modern software practices such as version-control and unit-testing. In this paper I present ```beamshapes```, a Python package that implements directivity patterns for sound source models. As of this publication, ```beamshapes``` version 0.2.0 implements four sound sources 1) piston in an infinite baffle, 2) point source of a sphere 3) oscillating cap of a sphere and 4) piston in a sphere. 
+A host of published directivity functions exist in the literature, but it is my experience that their computational implementations remain as inhouse scripts that are often in proprietary language platforms. To my knowledge there are no computational implementations of sound source directivities that are open-source and developed using modern software practices such as version-control and unit-testing (but see for larger packages [@pyroom;@levitate] with modules devoted to directivity implementations). In this paper I present ```beamshapes```, a Python package that implements directivity patterns for sound source models. As of this publication, ```beamshapes``` version 0.2.0 implements four sound sources 1) piston in an infinite baffle, 2) point source of a sphere 3) oscillating cap of a sphere and 4) piston in a sphere. 
 
 Computational implementations of directivity functions often require long run-times due to the intensive numerical routines and arbitrary precision math involved. Long run-times hinder scientific projects in reducing the number of models and parameter space that can be explored. ```beamshapes``` boasts parallelised code to generate significant speed-ups in run-times. 
 
