+++ paper.bib
@@ -27,6 +27,24 @@
   doi={10.1121/1.1912304}
 }
 
+@misc{levitate,
+  author = {Andersson, Carl},
+  title = {Levitate},
+  year = {2018},
+  publisher = {GitHub},
+  journal = {GitHub repository},
+  howpublished = {\url{https://github.com/AppliedAcousticsChalmers/levitate}},
+  commit = {c6ad1038327abfc82a5402b42019b69e52f7f5c4}
+}
+
+@inproceedings{pyroom,
+  title={Pyroomacoustics: A python package for audio room simulation and array processing algorithms},
+  author={Scheibler, Robin and Bezzam, Eric and Dokmani{\'c}, Ivan},
+  booktitle={2018 IEEE International Conference on Acoustics, Speech and Signal Processing (ICASSP)},
+  pages={351--355},
+  year={2018},
+  organization={IEEE}
+}
 
 @article{mogensen1979sound,
   title={Sound radiation patterns in the frequency domain of cries from a Vespertilionid bat},
