### 2021-10-21 7e7d75b
add separate example to showcase python-flint. skip travis;
