+++ beamshapes/piston_in_sphere.py
@@ -35,6 +35,7 @@ from joblib import Parallel, delayed
 import mpmath
 from mpmath import mpf
 import numpy as np
+import platform
 import sympy
 from sympy import symbols, legendre, sin, cos, tan, Sum, I, diff, pi, sqrt
 from sympy import Matrix, Piecewise
@@ -232,7 +233,12 @@ def compute_Mmn_parallel(params):
             
     multi_paramset_str = [args_to_str(**each) for each in multi_paramsets]
     num_cores = int(params.get('num_cores',-1))
-    M_mn_out = Parallel(n_jobs=num_cores, backend='multiprocessing')(delayed(parallel_calc_one_Mmn_term)(**inputs) for inputs in tqdm.tqdm(multi_paramset_str))
+    if platform.system()=='Linux':
+        M_mn_out = Parallel(n_jobs=num_cores, backend='multiprocessing')(delayed(parallel_calc_one_Mmn_term)(**inputs) for inputs in tqdm.tqdm(multi_paramset_str))
+    elif platform.system()=='Windows':
+        M_mn_out = Parallel(n_jobs=num_cores)(delayed(parallel_calc_one_Mmn_term)(**inputs) for inputs in tqdm.tqdm(multi_paramset_str))
+    else:
+        M_mn_out = Parallel(n_jobs=num_cores, backend='multiprocessing')(delayed(parallel_calc_one_Mmn_term)(**inputs) for inputs in tqdm.tqdm(multi_paramset_str))
     M_matrix = format_Mmn_to_matrix(M_mn_out)
     return M_matrix
 #%%
@@ -389,54 +395,54 @@ def piston_in_sphere_directivity(angles, params, **kwargs):
     return A_n, directivity
 
 
-# if __name__ == '__main__':
-#     import matplotlib.pyplot as plt
-#     frequency = mpmath.mpf(50*10**3) # kHz
-#     vsound = mpmath.mpf(330) # m/s
-#     wavelength = vsound/frequency
-#     alpha_value = mpmath.pi/3 # 60 degrees --> pi/3
-#     k_value = 2*mpmath.pi/(wavelength)
-#     ka_val = 1
-#     print(f'Starting piston in sphere for ka={ka_val}')
-#     ka = mpmath.mpf(ka_val)
-#     a_value = ka/k_value 
-#     R_value = a_value/mpmath.sin(alpha_value)  # m
-#     paramv = {}
-#     paramv['R'] = R_value
-#     paramv['alpha'] = alpha_value
-#     paramv['k'] = k_value
-#     paramv['a'] = a_value
+if __name__ == '__main__':
+    import matplotlib.pyplot as plt
+    frequency = mpmath.mpf(50*10**3) # kHz
+    vsound = mpmath.mpf(330) # m/s
+    wavelength = vsound/frequency
+    alpha_value = mpmath.pi/3 # 60 degrees --> pi/3
+    k_value = 2*mpmath.pi/(wavelength)
+    ka_val = 5
+    print(f'Starting piston in sphere for ka={ka_val}')
+    ka = mpmath.mpf(ka_val)
+    a_value = ka/k_value 
+    R_value = a_value/mpmath.sin(alpha_value)  # m
+    paramv = {}
+    paramv['R'] = R_value
+    paramv['alpha'] = alpha_value
+    paramv['k'] = k_value
+    paramv['a'] = a_value
     
     
-#     import pandas as pd
-#     df = pd.read_csv('workshop/ka5_piston_in_sphere.csv')
-#     df2 = pd.read_csv('tests/plots_data/piston_in_sphere_fig12-23.csv')
-#     ka5 = df2[df2['ka']==ka_val]
+    import pandas as pd
[…]
