### 2021-10-22 244dea4
change 4 models at a time example to be Windows compatible. closes #20
