+++ examples/0_preprint_figures.py
@@ -22,182 +22,186 @@ from beamshapes import piston_in_sphere_directivity
 from beamshapes import cap_in_sphere_directivity
 from beamshapes import point_source_on_a_sphere_directivity
 
-#%% set up the infinite baffle directivity on the left and in sphere on the right
-ka_values  = [1, 5, 10]
-a = 0.01 # piston of 1cm radius 
+# the 'if__name__' pattern for a module is odd - but is required when running code with 
+# the multiprocessing package in Windows. The example should work without the if __name__
+# in Linux systems.
+if __name__ == '__main__':
+    #%% Set the common parameters to make all source models comparable
+    ka_values  = [1, 5, 10]
+    a = 0.01 # piston of 1cm radius 
+    R = 0.1 # radius of a sphere
+    alpha = mpmath.pi/6
+    a_value = R*mpmath.sin(alpha)
+    angles = np.linspace(0,-np.pi,100)
+    
+    #%% set up the infinite baffle directivity 
+    piston_infbaf_D = np.zeros((angles.size, len(ka_values)))
+    
+    print('Calculating piston in an infinite baffle....')
+    for j,each_ka in enumerate(ka_values):
+        k_value = each_ka/a
+        parameters = {'k':k_value, 'a':a}
+        _, piston_infbaf_D[:,j] = piston_infbaf.piston_in_infinite_baffle_directivity(angles,
+                                                                                   parameters)
 
+    #%% 
+    # Now calculate the piston in sphere 
+    
+    
+    piston_in_sphere_D = np.zeros((angles.size, len(ka_values)))
+    
+    print('Calculating piston in a sphere....')
+    for j,each_ka in enumerate(ka_values):
+        k_value = each_ka/a_value
+        parameters = {'k':k_value, 'a':a_value, 'R': R, 'alpha':alpha}
+        A_n, piston_in_sphere_D[:,j] = piston_in_sphere_directivity(angles,
+                                                                        parameters)
 
-angles = np.linspace(0,-np.pi,100)
-piston_infbaf_D = np.zeros((angles.size, len(ka_values)))
-
-
-
-for j,each_ka in enumerate(ka_values):
-    k_value = each_ka/a
-    parameters = {'k':k_value, 'a':a}
-    _, piston_infbaf_D[:,j] = piston_infbaf.piston_in_infinite_baffle_directivity(angles,
-                                                                               parameters)
-
-#%% 
-# Now calculate the piston in sphere 
-
-
-R = 0.1 # radius of a sphere
-alpha = mpmath.pi/6
-a_value = R*mpmath.sin(alpha)
-piston_in_sphere_D = np.zeros(angles.size)
-
-for j,each_ka in enumerate(ka_values):
-    k_value = each_ka/a_value
-    parameters = {'k':k_value, 'a':a_value, 'R': R, 'alpha':alpha}
-    A_n, piston_in_sphere_D[:,j] = piston_in_sphere_directivity(angles,
-                                                                    parameters)
-
-
-
-#%% 
-# Now calculate oscillating cap of a sphere
-
-R = mpmath.mpf(0.1) # radius of a sphere
-cap_of_sphere = np.zeros(piston_infbaf_D.shape)
-
-for j,each_ka in enumerate(ka_values):
-    k_value = each_ka/a_value
-    parameters = {'k':k_value, 'a':a_value, 'R': R, 'alpha':alpha}
[…]
