### 2021-10-22 thejasvibr/bat_beamshapes#6 @thejasvibr
This error was caused by an implementation of the piston in a sphere using the ```python-flint``` package, instead of the  ``` mpmath ``` package. The ```python-flint``` package is *extremely* fast (in comparison to ```mpmath```), but is tricky to install and  currently only available on Linux systems. In hindsight, perhaps it's not the most welcoming example - requiring users to go through a long installation.

I've now done two things: 1) the 4 source model example does not require external dependency installation 2) made a separate example for more invested users who'd like to run the faster ```python-flint``` based implemention.
