### 2021-10-22 87b408d
closes #6; closes #13; skip travis;
1) Make the 4 model plot wihtout using flint.
2) Make examples Windows runnable.
