+++ examples/0_preprint_figures.py
@@ -13,6 +13,11 @@ Important to note
 As you may notice, some of the models require numerical inputs of a particular format (mpmath float for instance). 
 This is important to pay attention to, and may otherwise lead to avoidable errors. 
 
+Lengthy computations ahead
+~~~~~~~~~~~~~~~~~~~~~~~~~~
+Some of the models are computationally intensive and do take a long time, depending
+on your system. The piston in a sphere can take over 10 minutes to run for example!!
+
 """
 import mpmath
 import numpy as np 
@@ -26,7 +31,7 @@ from beamshapes import point_source_on_a_sphere_directivity
 # the multiprocessing package in Windows. The example should work without the if __name__
 # in Linux systems.
 if __name__ == '__main__':
-    #%% Set the common parameters to make all source models comparable
+#%% Set the common parameters to make all source models comparable
     ka_values  = [1, 5, 10]
     a = 0.01 # piston of 1cm radius 
     R = 0.1 # radius of a sphere
@@ -34,7 +39,7 @@ if __name__ == '__main__':
     a_value = R*mpmath.sin(alpha)
     angles = np.linspace(0,-np.pi,100)
     
-    #%% set up the infinite baffle directivity 
+#%% set up the infinite baffle directivity 
     piston_infbaf_D = np.zeros((angles.size, len(ka_values)))
     
     print('Calculating piston in an infinite baffle....')
@@ -44,8 +49,8 @@ if __name__ == '__main__':
         _, piston_infbaf_D[:,j] = piston_infbaf.piston_in_infinite_baffle_directivity(angles,
                                                                                    parameters)
 
-    #%% 
-    # Now calculate the piston in sphere 
+#%% 
+# Now calculate the piston in sphere 
     
     
     piston_in_sphere_D = np.zeros((angles.size, len(ka_values)))
@@ -57,9 +62,9 @@ if __name__ == '__main__':
         A_n, piston_in_sphere_D[:,j] = piston_in_sphere_directivity(angles,
                                                                         parameters)
 
-    #%% 
-    # Now calculate oscillating cap of a sphere
-    
+#%% 
+# Now calculate oscillating cap of a sphere
+
     R = mpmath.mpf(0.1) # radius of a sphere
     cap_of_sphere = np.zeros(piston_infbaf_D.shape)
     
@@ -71,8 +76,8 @@ if __name__ == '__main__':
                                                                         parameters)
     
     
-    #%% 
-    # Calculate the beamshape of a point on a sphere
+#%% 
+# Calculate the beamshape of a point on a sphere
     R = 0.1 
     point_on_sphere = np.zeros(piston_infbaf_D.shape)
     
@@ -82,7 +87,7 @@ if __name__ == '__main__':
         parameters = {'k':k_value, 'R':R}
         _, point_on_sphere[:,j] = point_source_on_a_sphere_directivity(angles, parameters)
     
-    #%%
+#%% Make the plot with all model results:
     
     def set_ylim():
         plt.ylim(-52,0)
@@ -107,10 +112,7 @@ if __name__ == '__main__':
     
     plt.xticks(np.radians(tick_angles) ,tick_labels , fontsize=6)
     a0.xaxis.set_tick_params(pad=-3.5)
-    #plt.legend(loc=(0.95,-0.07), frameon=False)
-    #plt.text(0.05, 1.02, 'piston in a sphere', transform=a0.transAxes, fontsize=8)
-    # plt.text(0.55, 1.02, 'piston in an infinite baffle', transform=a0.transAxes,
-    #                          multialignment='right', fontsize=8)
+    
[…]

[…]
