### 2021-10-22 thejasvibr/bat_beamshapes#13 @thejasvibr
Resolved. Examples run without external installation required. Please see explanation in #6
