### 2021-10-24 e369b65
skip travis; clean up docs, API portions and improve docs on the source models
