+++ beamshapes/piston_in_sphere.py
@@ -55,7 +55,7 @@ from beamshapes.utilities import args_to_mpmath, args_to_str
 r1 = (R*cos(alpha))/cos(theta)
 
 
-#%%
+
 # equation 12.106
 alternate_hankels = n*sph_hankel2.subs({'n':n-1, 'z':k*r1})-(n+1)*sph_hankel2.subs({'n':n+1, 'z':k*r1})
 pt1_postterm = legendre(n,cos(theta))*cos(theta)
@@ -72,7 +72,7 @@ Imn = Integral(Imn_term,(theta,0,alpha))
 Imn_term_func = lambdify([m,n,k,R,alpha, theta], Imn_term, 'mpmath')
 # Imn_func = lambdify([m,n,k,R,alpha],Imn,'mpmath') 
 
-#%%
+
 def Imn_func(mv,nv,kv,Rv,alphav):
     '''
     eqn. 12.106
@@ -83,7 +83,7 @@ def Imn_func(mv,nv,kv,Rv,alphav):
                        (0,alphav),
                        method='gauss-legendre')   
 
-#%%
+
 # equation 12.107
 Kmn_expr = legendre(n, cos(theta))*legendre(m, cos(theta))*sin(theta) # the integrl of this expression 
 # has a solution given in Appendix II, eqn 70
@@ -109,7 +109,7 @@ Kmn = Piecewise((eqn70_mnoteqn,m>n),
                 (eqn70_meqn,True), )
 Kmn_func = lambdify([m,n,alpha],Kmn,'mpmath')
 
-#%%
+
 # equation 12.108
 Lm_expr = legendre(m,cos(theta))*(r1**2/R**2)*tan(theta)
 Lm = Integral(Lm_expr, (theta,0,alpha))
@@ -135,13 +135,13 @@ def Lm_func(mv,Rv,alphav):
 #     errors.append(err)
 # print(errors)
 # # what does the Lm term plot like? 
-# #%% 
+#  
 # lm_theta = lambda mv, thetav: Lm_term(mv, paramv['R'],paramv['alpha'], thetav)
 # thet = mpmath.linspace(0, mpmath.pi/3,50)
 # plt.figure()
 # plt.plot(np.degrees(np.float32(thet)), [lm_theta(25, each) for each in thet])
 
-#%%
+
 # b matrix
 b = -I*Lm
 # b_func = lambdify([m,alpha], b,'mpmath') # eqn. 12.104
@@ -154,7 +154,7 @@ def b_func(mv, Rv, alphav):
     Lm_value = Lm_func(mv,Rv,alphav) # Integral wrt theta
     return -mpmath.j*Lm_value
 
-#%% 
+ 
 # Setting up the matrices 
 # %% 
 mmn_hankels = n*sph_hankel2.subs({'n':n-1,'z':k*R})-(n+1)*sph_hankel2.subs({'n':n+1,'z':k*R})
@@ -178,7 +178,7 @@ def calc_N(params):
 
 ### parallel zone 
 def calc_one_Mmn_term(**params):
-    '''
+    '''Calculates on Mmn term given the input m,n,k,R and alpha terms.
     '''
     Imn_value = Imn_func(params['m'], params['n'],params['k'],
                          params['R'],params['alpha'])
@@ -189,6 +189,8 @@ def calc_one_Mmn_term(**params):
     return numerator/denom
 
 def parallel_calc_one_Mmn_term(**args):
+    '''Parallelised version of `calc_one_Mmn_term`
+    '''
     mpmath_args = args_to_mpmath(**args)
     output = calc_one_Mmn_term(**mpmath_args)
     backto_str = str(output)
@@ -236,16 +238,17 @@ def compute_Mmn_parallel(params):
[…]
