+++ docs/source/source_models.rst
@@ -2,17 +2,17 @@ Source Models API
 =================
 
 .. automodule:: beamshapes.piston_in_infinite_baffle
-       :members:
+       :members: d_theta_func, piston_in_infinite_baffle_directivity
 
 
 .. automodule:: beamshapes.point_source_on_a_sphere
-       :members:
+       :members: point_source_on_a_sphere_directivity, d_zero_func, d_theta
 
 .. automodule:: beamshapes.cap_in_sphere
-       :members:
+       :members: cap_in_sphere_directivity, relative_directivity_db, d_zero, d_theta
 
 .. automodule:: beamshapes.piston_in_sphere
-       :members:
+       :members: d_theta, d_zero, relative_directivity_db, piston_in_sphere_directivity
 
 .. automodule:: beamshapes.piston_in_sphere_flint
-       :members:
+       :members: make_Mmn_pll, dtheta, dzero, directivity

[…]
