### 2021-10-25 e50d7b3
more cleaning of docs [skip travis]
