+++ docs/source/general_intro.rst
@@ -4,23 +4,6 @@ Sound sources don't radiate sound uniformly most of the time. Even us humans for
 
 Let's go through the main concepts required to use the `beamshapes` package.
 
-
-Input Parameters
-----------------
-The inputs will tend to be model-specific, but the common input parameters
-to keep in mind are:
-
-    #. `k`, wavenumber. This is :math:`\frac{2\pi}{\lambda}`
-    #. :math:`\lambda`, wavelength or :math:`\frac{v_\text{sound}}{\text{frequency}}`
-    #. `a` : piston radius, wherever applicable
-    #. `R` : sphere radius, wherever applicable. 
-    #. :math:`\alpha, \theta, \phi`: various angles describing the size of the oscillating portion. These angles are in radians - not degrees!
-
-The values of these input parameters are stored in a dictionary.
-Remember to try and use :code:`mpmath` objects (eg. :code:`mpmath.pi`, or :code:`mpmath.mpf`)
-to set the input parameter values if you're interested in getting  high-precision output.      
-This means, eg. using :code:`mpmath.pi/3` instead of :code:`np.pi/3`, or :code:`math.pi/3`.
-
 Source models
 -------------
 A 'beamshape' is tightly bound to the input parameters, and the sound source model behind
@@ -76,11 +59,15 @@ A sphere is sliced, and the 'cap' is discarded. The 'open' portion of the sliced
 This piston in the sphere vibrates to produce sound. The parameters needed to define this model are wavenumber, sphere radius, 
 aperture angle of the piston, and piston radius. 
 
+Input Parameters
+----------------
+The inputs will tend to be model-specific, but the common input parameters
+to keep in mind are:
 
-
-
-
-
+    #. `k`, wavenumber. This is :math:`\frac{2\pi}{\lambda}` - this is another wave of defining the frequency of the vibrating part, and the sound produced. :math:`\lambda`is the wavelength of the sound, also defined as :math:`\frac{v_\text{sound}}{\text{frequency}}`. 
+    #. `a` : piston radius, wherever applicable
+    #. `R` : sphere radius, wherever applicable. 
+    #. :math:`\alpha, \theta, \phi`: various angles describing the size of the oscillating portion. These angles are in radians - not degrees!
 
 
 Directionality functions
@@ -105,14 +92,14 @@ In case the chosen source model's directionality requires intensive calculations
 to generate estimates, then the calculated outputs are also returned. 
 
 .. code-block:: shell
-    
+
     # .... having chosen certain input parameters and put them into input_params 
     >>> import mpmath
     >>> import beamshapes as beamshapes
     >>> from beamshapes import piston_in_sphere_directionality
     >>> angles = mpmath.linspace(0,pi,10) 
     >>> An_out, spherepiston_beam = piston_in_sphere_directionality(angles, input_params)
-    # in case of wanting to calculate more points now - it saves time to do this:
+    # in case you need to calculate more points now - it saves time to do this:
     >>> input_params['An'] = An_out
     >>> new_angles = mpmath.linspace(0,pi,100)
     >>> _, detailed_spherepiston_beam = piston_in_sphere_directionality(new_angles, input_params)
