### 2021-10-25 thejasvibr/bat_beamshapes#15 @thejasvibr
Resolved in [d3248f](https://github.com/thejasvibr/bat_beamshapes/commit/d3248f4c15913288bd4ea00ba79d8bdc139b3070) . Example with ```python-flint``` dependency run in a separate module (and verified to run on Linux system).

Docs now compiling with minor warnings (that can be ignored).
