### 2021-10-25 d690fad
clean up docs a bit and set thumbnails for examples
