+++ docs/source/general_intro.rst
@@ -1,16 +1,23 @@
-General Intro to `beamshapes`
-=============================
+Concepts behind :code:`beamshapes`
+==================================
 Sound sources don't radiate sound uniformly most of the time. Even us humans for instance, we emit more energy while talking to the front than to the back. The exact (non-uniform) pattern in which sound is radiated defines the 'beamshape' of a source. The beamshape is typically a combination of the frequency of emitted sound and the geometry of the vibrating surface and its associated (non-vibrating) surfaces. 
 
 Let's go through the main concepts required to use the `beamshapes` package.
 
 Source models
 -------------
-A 'beamshape' is tightly bound to the input parameters, and the sound source model behind
-it eg. piston in a sphere, cap of a sphere etc. The source model refers to the source
-of sound and its geometric properties, and assumptions of the vibrations etc. According
+The source model refers to the source of sound and its geometric properties, and assumptions of the vibrations etc. According
 to the situation in hand, different source models may be physically/biologically relevant! 
 
+
+There are two main ways to predict how sound will radiate from a source - analytical or numerical modelling. Analytical models
+start with equations defining the physics of sound radiation and move on to produce mathematical solutions. Numerical
+methods use various numerical algorithms (eg. finite-element method) to computationally simulate sound radiation. :code:`beamshapes`
+specifically implements analytical source models with pre-defined solutions. The advantage of using such analytical models is the lower
+number of parameters needed to describe and understand the resulting sound radation. 
+
+Below is a brief description of the source models and the parameters relevant to them. For more information on each of the model's please refer to 
+the source references. 
 
 Piston in an infinite baffle
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~	
@@ -22,6 +29,8 @@ Piston in an infinite baffle
 Here a rigid circular disk (the 'piston') vibrates back and forth between a hole in a huge baffle that matches the disk's size.
 The parameters needed to define this model are the wavenumber and piston radius. 
 
+Reference : Chp 13, Beranek, L. L., & Mellow, T. (2019). Acoustics: sound fields and transducers. Academic Press.
+
 Point on a sphere
 ~~~~~~~~~~~~~~~~~
 .. image:: ./_static/point_on_sphere.png.png
@@ -32,6 +41,8 @@ Point on a sphere
 An infinitesimally small portion of a sphere's surface (the 'point') is considered to vibrate. The rest of the sphere does not vibrate.
 The parameters needed to define this model are the wavenumber and sphere radius. 
 
+Reference: Chp 12, Beranek, L. L., & Mellow, T. (2019). Acoustics: sound fields and transducers.
+Academic Press.
 
 Oscillating cap of a sphere
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~
@@ -47,6 +58,9 @@ relation :math:`u(R,\theta) = u_{0}cos \theta`, where :math:`\theta` is the dist
 
 The parameters needed to define this model are wavenumber, sphere radius, and the aperture angle of the cap. 
 
+Reference: Chp 12, Beranek, L. L., & Mellow, T. (2019). Acoustics: sound fields and transducers.
+Academic Press.
+
 Piston in a sphere
 ~~~~~~~~~~~~~~~~~~
 
@@ -59,51 +73,18 @@ A sphere is sliced, and the 'cap' is discarded. The 'open' portion of the sliced
 This piston in the sphere vibrates to produce sound. The parameters needed to define this model are wavenumber, sphere radius, 
 aperture angle of the piston, and piston radius. 
 
+Reference: Chp 12, Beranek, L. L., & Mellow, T. (2019). Acoustics: sound fields and transducers.
+Academic Press.
+
[…]
