+++ examples/README.rst
@@ -1,3 +1,5 @@
 Examples
 ========
-A set of examples showing the capabilities  of the `beamshapes` package. 
+A set of examples showing the capabilities  of the :code:`beamshapes` package. 
+
+To get an overview of how to use `all` the currently implemented packages, start with `4 models at a time`. 
