### 2021-10-25 52d8021
[skip travis] clean up docs even more
