+++ examples/pistoon_inf_baffle.py
@@ -9,7 +9,8 @@ Piston in an infinite baffle example
 #%% The piston in an infinite baffle is a classic model used to describe
 # the beamshapes of mouth-emitting bat species [1,2] (among other animals eg. [3]). 
 # It's easily to compute, but however can only describe the beamshape upto :math:`\pm90^{\circ}`
-# off-axis. Let's see an example of what the piston model outputs:
+# off-axis. One major issue with piston in an infinite baffle is that it is front-back symmetric,
+# which is unrealistic for most animal vocalisations. Let's see an example of what the piston model outputs:
 
 
 from beamshapes import piston_in_infinite_baffle_directivity
@@ -22,7 +23,7 @@ import numpy as np
 
 dB = lambda X: 20*np.log10(abs(X))
 
-angles = np.linspace(0,np.pi/2,100)
+angles = np.linspace(0,np.pi,100)
 k = 10.0
 ka_values = np.array([1,3,5,10])
 a_values = ka_values/k
@@ -31,14 +32,14 @@ parameters = {'k':k}
 
 plt.figure()
 a0 = plt.subplot(111, projection='polar')
-
+a0.set_theta_zero_location('N')
 for ka_v, a_v in zip(ka_values, a_values):
     parameters['a'] = a_v
     _, dirn = piston_in_infinite_baffle_directivity(angles, parameters)
     plt.plot(angles, np.array(dirn), label=str(ka_v))
     angles *= -1 # switch between L & R of 
 plt.legend(title='ka')
-plt.savefig('PIB.png')
+#plt.savefig('../docs/source/_static/PIB.png')
 
 
 #%%
