### 2021-10-25 thejasvibr/bat_beamshapes#7 @thejasvibr
As of [52d8021](https://github.com/thejasvibr/bat_beamshapes/commit/52d80210570881d3169eba4c83b1b75bd3b70cad) Have worked on adding brief model descriptions and background on their meaning.

Hope this version of the docs is clearer and provides more background?
