### 2021-10-25 thejasvibr/bat_beamshapes#5 @thejasvibr
I haven't found a way to render the ```lambdify``` output as math. Even if the math rendering did work out - it'd be a lot of math equations getting in the way of the broad picture of how the package works from a software point?

I reconsidered what is relevant enough to be on the online API, and have thus trimmed the functions whose documentations will be hosted online.

Hope this version (as of [52d8021](https://github.com/thejasvibr/bat_beamshapes/commit/52d80210570881d3169eba4c83b1b75bd3b70cad)) of the  API docs is clearer?
