### 2021-10-25 6fa5640
clean up commented out portions of code all through the code base [skip travis]
closes #3
