+++ beamshapes/sim_reclevels.py
@@ -151,51 +151,3 @@ def sim_miclevel_beamshape(sl, call_direction,
     
     miclevel_beamshape = mic_levels_omni + rel_onaxis
     return miclevel_beamshape
-
-# if __name__=='__main__':
-#     print('yay')
-#     mic_positions = np.array([[-1,1],
-#                               [0,1],
-#                               [1,1],
-#                               [1.5,1.5],
-#                               [2,0.75],
-#                               [-1,-1]])
-#     bat_position = np.array([0,0.0])
-#     call_dirn = np.radians(225)
-#     angles = angle_bet_bat_and_mic(bat_position, mic_positions, call_dirn)
-#     print(np.round(np.degrees(angles),2))
-#     #%%
-#     plt.figure()
-#     plt.plot(bat_position[0],bat_position[1],'*')
-#     plt.plot(mic_positions[:,0],mic_positions[:,1],'*')
-#     for num,each in enumerate(mic_positions):
-#         plt.annotate(str(num),each)
-        
-#     r = 0.1
-#     plt.arrow(bat_position[0],bat_position[1],
-#                           r*np.cos(call_dirn),
-#                           r*np.sin(call_dirn),width=0.01)
-    
-#     #%% Showing the effect of the asymmetry parameter
-#     thetas = np.linspace(0,2*np.pi,100)
-#     plt.figure()
-#     a0 = plt.subplot(111, projection='polar')
-#     a0.set_theta_zero_location('N')
-#     a0.set_theta_direction(1)
-#     plt.plot(thetas, simple_cardiod(thetas,A=5),label='A=5')
-#     plt.plot(thetas, simple_cardiod(thetas,A=10),label='A=10')
-#     plt.plot(thetas, simple_cardiod(thetas,A=20),label='A=20')
-#     plt.legend()
-
-#     # %% 
-#     rel = simple_cardiod(thetas,A=20)
-#     rel_x = np.cos(thetas)
-#     rel_y = np.sin(rel)
-#     plt.figure()
-#     plt.plot(thetas, rel)
-#     #%% with cardiod pattern
-#     calldirn = 0
-#     op = sim_miclevel_beamshape(100, calldirn, bat_position, 
-#                            mic_positions,beamshape_model='cardiod',
-#                            model_params={'A':7.3})
-#     print(op)
