### 2021-10-25 thejasvibr/bat_beamshapes#3 @thejasvibr
#%% symbols and commented out code portions deleted.
