### 2021-10-27 thejasvibr/bat_beamshapes#8 @thejasvibr
Hi @nils-werner , thanks for your comment on the tests.
Admittedly the tests in the repo are do have some boilerplate related to loading the ground-truth data and testing each parameter set - and would need time for a future contributor (or myself) to get into.

About switching to ```pytest```  and re-factoring the tests, I have considered these options, and finally realised my current project + funding timeline means it won't be possible to achieve within the next few months.

I do see the power of using a another testing framework, and the importance of having minimal boilerplate - and will consider  implementing these over a longer time-scale when possible. Hope this is okay?
