### 2021-10-28 thejasvibr/bat_beamshapes#9 @hagenw
Thanks for adding this.

The formatting is slightly broken:

![image](https://user-images.githubusercontent.com/173624/139208571-f82d788e-493e-40c7-8d25-4ddc6a366aca.png)
