### 2021-10-28 thejasvibr/bat_beamshapes#19 @hagenw
Thanks for mentioning them. It would maybe help to add a few more words on comparing them. I also do not really know/have used the other packages. This means a user I would have no idea when I should go for `beamshapes` or `pyroomacoustics` when reading the paper.
