### 2021-10-28 thejasvibr/bat_beamshapes#15 @hagenw
When I repeat my commands from above, I'm getting the following error:

```
WARNING: autodoc: failed to import module 'piston_in_sphere_flint' from module 'beamshapes'; the following exception was raised:
Traceback (most recent call last):
  File "/home/audeering.local/hwierstorf/.envs/bat_beamshapes/lib/python3.6/site-packages/sphinx/ext/autodoc/importer.py", line 70, in import_module
    return importlib.import_module(modname)
  File "/usr/lib/python3.6/importlib/__init__.py", line 126, in import_module
  File "<frozen importlib._bootstrap>", line 994, in _gcd_import
  File "<frozen importlib._bootstrap>", line 971, in _find_and_load
  File "<frozen importlib._bootstrap>", line 955, in _find_and_load_unlocked
  File "<frozen importlib._bootstrap>", line 665, in _load_unlocked
  File "<frozen importlib._bootstrap_external>", line 678, in exec_module
  File "<frozen importlib._bootstrap>", line 219, in _call_with_frames_removed
  File "/home/audeering.local/hwierstorf/.envs/bat_beamshapes/lib/python3.6/site-packages/beamshapes-0.2.1-py3.6.egg/beamshapes/piston_in_sphere_flint.py", line 32, in <module>
    from beamshapes.flint_parallelisation import conv_acb_to_str, conv_str_acb, interchange_params_and_str
  File "/home/audeering.local/hwierstorf/.envs/bat_beamshapes/lib/python3.6/site-packages/beamshapes-0.2.1-py3.6.egg/beamshapes/flint_parallelisation.py", line 15, in <module>
    acb = flint.acb
AttributeError: module 'flint' has no attribute 'acb'
```

Does this mean I have to install an additional package to fix this?
