### 2021-10-28 thejasvibr/bat_beamshapes#5 @hagenw
If I understand it correctly you decided now against showing the math as it is complicated and will not help the user, correct?
I think this is fine as you provide references to the underlying math anyway.

There is still one function left in the API that looks as before:

![image](https://user-images.githubusercontent.com/173624/139212070-10bea503-a141-419b-a768-e131f50650fd.png)

Is this on purpose, or did you forgot to change it there?
