### 2021-10-28 thejasvibr/bat_beamshapes#12 @hagenw
As I only use `pytest` I cannot directly advice how to use `coverage` with `unittest`.
I followed the steps you mentioned in `CONTRIBUTING` and get:

```bash
$ python -m unittest
...
Ran 6 tests in 399.609s

OK

$ coverage report
No data to report.
```
So, it does not work directly. Do I have to do another step?
