### 2021-10-29 042f1db
include correct Travis CI badge
