+++ README.md
@@ -1,4 +1,4 @@
-[![Build Status](https://travis-ci.com/thejasvibr/bat_beamshapes.svg?branch=dev)](https://travis-ci.com/thejasvibr/bat_beamshapes)
+[![Build Status](https://app.travis-ci.com/thejasvibr/bat_beamshapes.svg?branch=dev)](https://app.travis-ci.com/thejasvibr/bat_beamshapes)
 [![status](https://joss.theoj.org/papers/820fd37f8255a8c533d6cc4c9475ecb5/status.svg)](https://joss.theoj.org/papers/820fd37f8255a8c533d6cc4c9475ecb5)
 [![Documentation Status](https://readthedocs.org/projects/beamshapes/badge/?version=latest)](https://beamshapes.readthedocs.io/en/latest/?badge=latest)
 
