### 2021-10-29 4bf3645
include additional step in quantifying test coverage [travis skip]
