### 2021-10-29 70da621
* add missing step to estimate test coverage. * include instructions to run single module unit tests
