+++ CONTRIBUTING.md
@@ -19,12 +19,15 @@ If you wish to contribute a directivity implementation for a source-model see th
 For reasons of code homogeneity, I request that you keep the template structure suggested in the [how-to](https://beamshapes.readthedocs.io/en/latest/general_workflow.html) page. All implementations must have tests that at least check the coded implementation matches the published results of the original paper/book that described the model. (If you are planning to implement a directivity that isn't published or have calculated yourself - we still need tests, but I'd be curious to hear how to check the validity of the implementation's results!)
 
 ##  Unit-tests
-All unit-tests in the ```beamshapes``` package are written using the ```unittest``` package.
+All unit-tests in the ```beamshapes``` package are written using the ```unittest``` package. Be aware that running all the tests will take a decent amount of time on an average PC/laptop (anywhere upwards of 20 minutes!).
 
 ### Running unit-tests
-To run the unittests, first clone and install the ````beamshapes``` package (see the [README](README.md)).To run the tests you will need to install all the packages in the ```tests/requirements_test.txt``` file. From the root of the directory, open your Python command line tool of choice and run:
+To run the unittests, first clone and install the ```beamshapes``` package (see the [README](README.md)).To run the tests you will need to install all the packages in the ```tests/requirements_test.txt``` file. From the root of the directory, open your Python command line tool of choice and run:
 
 ```python -m unittest```  
+
+To run specific unit tests, use the commande ```python -m unittest beamshapes/tests/{name-of-test-here.py}```, where *{name-of-test-here.py}* is the 
+test you want to run.
 
 ### Writing unit-tests
 The unit-tests in ```beamshapes``` should test the computational implementation for correctness based on previously published results. See [here](beamshapes/tests/testing.md) for more on the testing approach. 
