### 2021-10-29 thejasvibr/bat_beamshapes#12 @thejasvibr
There was a missing step, which was to run the unit-tests through the ```coverage``` package (``` coverage run -m unittest discover ```). Have added it now as [70da621](https://github.com/thejasvibr/bat_beamshapes/commit/70da621b228226961d71d6cf6115723fd50d196b).

Hope the coverage reports provide sensible output now?
