### 2021-10-29 thejasvibr/bat_beamshapes#12 @hagenw
```bash
$ coverage report
Name                                             Stmts   Miss  Cover
--------------------------------------------------------------------
beamshapes/__init__.py                               5      0   100%
beamshapes/cap_in_sphere.py                         76     34    55%
beamshapes/piston_in_infinite_baffle.py             18      0   100%
beamshapes/piston_in_sphere.py                     144     12    92%
beamshapes/piston_in_sphere_flint.py               168    165     2%
beamshapes/point_source_on_a_sphere.py              26      0   100%
beamshapes/special_functions.py                      9      0   100%
beamshapes/tests/__init__.py                         0      0   100%
beamshapes/tests/test_osccapsphere.py               41      3    93%
beamshapes/tests/test_pistoninfbaffle.py            32      1    97%
beamshapes/tests/test_pistoninsphere_mpmath.py      83      2    98%
beamshapes/tests/test_pistoninsphereflint.py        57     42    26%
beamshapes/tests/test_pointsourcesphere.py          49      1    98%
beamshapes/utilities.py                              8      1    88%
beamshapes/version.py                                1      0   100%
--------------------------------------------------------------------
TOTAL                                              717    261    64%
```
