### 2021-10-31 74636ab
actually add docstrings for lambdified SymPy expressions for point on a sphere. also build on docs where they were missing.
