+++ docs/source/source_models.rst
@@ -4,15 +4,14 @@ Source Models API
 .. automodule:: beamshapes.piston_in_infinite_baffle
        :members: d_theta_func, piston_in_infinite_baffle_directivity
 
-
 .. automodule:: beamshapes.point_source_on_a_sphere
-       :members: point_source_on_a_sphere_directivity, d_zero_func, d_theta
+       :members: point_source_on_a_sphere_directivity, d_zero_func, d_theta_func
 
 .. automodule:: beamshapes.cap_in_sphere
-       :members: cap_in_sphere_directivity, relative_directivity_db, d_zero, d_theta
+       :members: cap_in_sphere_directivity, d_zero, d_theta
 
 .. automodule:: beamshapes.piston_in_sphere
-       :members: d_theta, d_zero, relative_directivity_db, piston_in_sphere_directivity
+       :members: d_theta, d_zero, piston_in_sphere_directivity
 
 .. automodule:: beamshapes.piston_in_sphere_flint
-       :members: make_Mmn_pll, dtheta, dzero, directivity
+       :members: make_Mmn_pll, dtheta, dzero, piston_in_sphere_directivity
