### 2021-10-31 thejasvibr/bat_beamshapes#5 @thejasvibr
I had initially left the weird lambda formatting to keep consistency in which function's docs were displayed - I [have now fixed the formatting issue](https://beamshapes.readthedocs.io/en/latest/source_models.html#point-source-on-a-sphere).

I managed to solve the formatting issue by realising a lambdified expression is after all a lambda object. Thanks to [this SO contribution](https://stackoverflow.com/a/50371184/4955732), I could add docstrings by separately setting the lambda functions ```.__doc__``` attribute. The output is now 'standard' looking as of [this commit](https://github.com/thejasvibr/bat_beamshapes/commit/74636abfda0c043f4dceb2a6bf0f991343747172).
