### 2021-11-01 53b5d13
start writing parameter estimation example on pymc3eg branch
