### 2021-11-05 915637a
work on parameter estimation example -set up the simulated data for an omnidirectional call.
