### 2021-11-07 3cfd2d8
add example which makes simulated data for parameter estimation
