+++ examples/param-estimation.py
@@ -10,12 +10,28 @@ sound directionality and actually being able to estimate many other aspects of s
 Let's simulate an example to show the power of actually using a source model to study 
 animal sound production. Of course, any type of sound can be studied this way!! 
 
-'''
-import matplotlib.pyplot as plt
-import numpy as np 
-import beamshapes.sim_reclevels as simlevels
-import mpmath
+This example will go through the following steps:
 
+    1. Define mic and bat positions 
+    2. Define call direction, on-axis level and parameters of the source model
+    3. Calculate received levels at mics assuming an omnidirectional source
+    4. Calculate mic to call-direction relative angles 
+    5. Generate beamshape from source model 
+    6. Add/subtract the relative change for off-axis angles. 
+    
+
+'''
+# sphinx_gallery_thumbnail_path = '_static/simulated_data.png'
+
+#%% the if __name__ == '__main__' is required when running on Windows systems
+if __name__ == '__main__':
+    
+    import matplotlib.pyplot as plt
+    import numpy as np 
+    import pandas as pd
+    import beamshapes.piston_in_sphere
+    import beamshapes.sim_reclevels as simlevels
+    import mpmath
 
 #%% 
 # Let's simulate a bat that emits five calls as it flies through a room with 12 microphones. 
@@ -24,80 +40,123 @@ import mpmath
 # These two trends mimic what a bat coming in to capture an insect will do. 
 
 # Define mic positions  in 2D at flight height of the bat
-mic_posns = np.zeros((12,2))
-# define x-positions
-mic_posns[:4,0] = 0 # first three mics on left wall
-mic_posns[4:8,0] = np.linspace(0.5,2,4)
-mic_posns[-4:,0] = 2.5
-
-# define y-positions 
-mic_posns[:4,1] = np.linspace(0,2,4)
-mic_posns[4:8,1] = 2.5
-mic_posns[-4:,1] = np.linspace(0,2,4)[::-1]
-
-# Define on-axis level for each of the 5 calls (dB SPL re 20 muPa at 1m)
-# These levels are semi-realistic!
-onaxis_levels = [100, 96, 90, 84, 84]
-
-# Define the directivity function broadly by altering ka for each call
-overall_ka = [10, 5, 4, 3, 2]
+    mic_posns = np.zeros((12,2))
+    # define x-positions
+    mic_posns[:4,0] = 0 # first three mics on left wall
+    mic_posns[4:8,0] = np.linspace(0.5,2,4)
+    mic_posns[-4:,0] = 2.5
+    
+    # define y-positions 
+    mic_posns[:4,1] = np.linspace(0,2,4)
+    mic_posns[4:8,1] = 2.5
+    mic_posns[-4:,1] = np.linspace(0,2,4)[::-1]
+    
+    # Define on-axis level for each of the 5 calls (dB SPL re 20 muPa at 1m)
+    # These levels are semi-realistic!
+    onaxis_levels = [100, 96, 90, 84, 84]
+    
+    # Define the directivity function broadly by altering ka for each call
+    # Set 'k' for a sound of 80 kHz frequency
+    overall_ka = [5, 4, 3, 2, 1]
+    
 #%%
 # Where was the bat at each of the 5 call emissinos?
-flight_path_x =  np.array([0.45, 1.0, 1.5, 2.0, 2.4])
-flight_path_y = np.array([0.3, 0.8, 1.2, 0.6, 0.4])
-flight_path = np.column_stack((flight_path_x, flight_path_y))
-
-# Where was the bat aiming it's call? 
-call_directions = np.pi/2-np.deg2rad(np.array([15, 60, 140, 160, 200]))
-
-
-plt.figure()
-plt.plot(flight_path[:,0], flight_path[:,1], '*')
[…]
