### 2021-11-07 f3f9d9d
finalise simulated data generation and start work on parameter estimation example
