+++ examples/parameter_estimation_1.py
@@ -1,0 +1,10 @@
+# -*- coding: utf-8 -*-
+"""
+
+Source parameter estimation - 
+============================
+Created on Sun Nov  7 10:52:14 2021
+
+@author: theja
+"""
+
