### 2021-11-10 thejasvibr/bat_beamshapes#22 @hagenw
Return types in API documentation inconsistent?
Related to https://github.com/openjournals/joss-reviews/issues/3740

When looking at the API documentation at https://beamshapes.readthedocs.io/en/latest/source_models.html
you will find

![image](https://user-images.githubusercontent.com/173624/141095254-8e266ab5-b082-4baa-a784-68c0078e4e3c.png)

Does this really don't have any return value?

In addition, we have a few that return a tuple, but the first entry is specified as `None`

![image](https://user-images.githubusercontent.com/173624/141095380-bd47d4f2-d3d8-494f-a65f-309b0fe5f329.png)

is this the desired behavior?

For others the "Return Type" is explicitly given

![image](https://user-images.githubusercontent.com/173624/141095542-f5447eba-85ec-418f-88c0-edde93f1599d.png)

---

This all makes it hard for the reader to know what to expect from the functions without trying them out.
