### 2021-11-10 thejasvibr/bat_beamshapes#11 @hagenw
It's fixed in the latest version of the docs:

![image](https://user-images.githubusercontent.com/173624/141095784-1aabbaeb-b643-417b-a329-31824b09a997.png)
