### 2021-12-06 thejasvibr/bat_beamshapes#5 @faroit
@nils-werner can this be closed?
