### 2021-12-06 thejasvibr/bat_beamshapes#7 @faroit
@nils-werner are these changes sufficient?
