### 2021-12-06 thejasvibr/bat_beamshapes#11 @faroit
@hagenw can this be closed?
