### 2021-12-06 thejasvibr/bat_beamshapes#11 @hagenw
Yes, but I don't have the rights to close it.
