### 2021-12-06 thejasvibr/bat_beamshapes#11 @hagenw
I have the rights, sorry ;)
