### 2021-12-06 thejasvibr/bat_beamshapes#19 @hagenw
I would like to reopen this issue, as I don't think the current state in the paper is sufficient.
