### 2021-12-06 #3740 @hagenw
There is only one issue left for me: https://github.com/thejasvibr/bat_beamshapes/issues/19
It's officially closed, but I do not think that the solution is sufficient.
