### 2021-12-06 thejasvibr/bat_beamshapes#19 @faroit
@thejasvibr would you be willing to give this another pass?
