### 2021-12-07 thejasvibr/bat_beamshapes#19 @thejasvibr
hello @faroit @hagenw - will give another try to make this portion clearer. Apologies for the delay in response - was on leave till today.
