### 2021-12-07 a2b5840
[skip travis] include portion in manuscript placing beamshapes wrt other packages computing directivities.
