+++ paper.md
@@ -33,7 +33,7 @@ functions exist for a wide variety of sound sources that can be modelled analyti
 of a directivity function is of the piston in an infinite baffle. The piston is a circular surface of radius
 $a$, vibrating back and forth about a hole in an infinite wall or baffle. The directivity is described by $\frac{2J_{1}(ka \times sin \theta)}{ka \times sin \theta}$, where $J_{1}$ is the Bessel's function of the first kind, $k$ is the wavenumber, where $k = \frac{2\pi f}{c}$, where $f$ is the frequency of the sound, and $c$ is the speed of sound. The angle of the receiver is $\theta$, which varies from from 0-2$\pi$ radians in the azimuth. 
 
-Directivity functions can be used in a two-fold manner 1) to deliberately engineer devices to suit particular specifications, eg. loudspeaker sound fields [@beranek2012acoustics] and 2) to infer parameters of a sound source itself having assumed a relevant model, eg. estimating direction of call emission [@guarato2011method] and mouth aperture of bat echolocation calls [@jakobsen2013convergent;@kounitsky2015bats].
+Directivity functions can be used in a two-fold manner 1) to deliberately engineer devices to suit particular specifications, e.g. loudspeaker sound fields [@beranek2012acoustics] and 2) to infer parameters of a sound source itself having assumed a relevant model, e.g. estimating direction of call emission [@guarato2011method] and mouth aperture of bat echolocation calls [@jakobsen2013convergent;@kounitsky2015bats].
 
 
 ![Directivity patterns ($\frac{D_{\theta}}{D_{0}}$) of the currently implemented sound sources for a common set of $ka$ values for comparison, where $k$ is the wavenumber ($\frac{2\pi\:f}{c}$) and $a$ is the piston radius - or its equivalent. The directivity pattern shows the ratio between the off-axis sound level at ($\theta^{\circ}$) to the on-axis level at ($0^{\circ}$) in decibels. A) piston in an infinite baffle B) piston in a sphere with the half-angle aperture $\alpha = 30^{\circ}$, and where the piston radius $a=R\:sin \:\alpha$ C) oscillating cap of a sphere with  $\alpha=30^{\circ}$, and the equivalent piston radius is $a=R\:sin \:\alpha$ D) vibrating point on a sphere, here the $kR=ka$ for comparison with the other models. \label{fig:pistonsphereinf}](paper_related/piston_sphere_baffle.png)
@@ -41,11 +41,11 @@ Directivity functions can be used in a two-fold manner 1) to deliberately engine
 
 # Statement of need
 
-A host of published directivity functions exist in the literature, but it is my experience that their computational implementations remain as inhouse scripts that are often in proprietary language platforms. To my knowledge there are no computational implementations of sound source directivities that are open-source and developed using modern software practices such as version-control and unit-testing (but see for larger packages [@pyroom;@levitate] with modules devoted to directivity implementations). In this paper I present ```beamshapes```, a Python package that implements directivity patterns for sound source models. As of this publication, ```beamshapes``` version 0.2.0 implements four sound sources 1) piston in an infinite baffle, 2) point source of a sphere 3) oscillating cap of a sphere and 4) piston in a sphere. 
[…]
