### 2021-12-07 thejasvibr/bat_beamshapes#19 @thejasvibr
Have included clarification in the MS as of [a2b5840...](https://github.com/thejasvibr/bat_beamshapes/commit/a2b5840d863b2571ae29f742beea73873581a36e).

To be specific, I've actually only focussed on the ```levitate``` package as I realised that ```pyroomacoustics```  actually implements the cardiod pattern - which describes a 'shape' rather than a directivity pattern linked to a specific sound source model. ```levitate``` implements relatively 'simple' sound source directivities with analytical solutions - while ```beamshapes``` stands out in implementing directivities that also don't have analytical solutions.
