### 2021-12-07 ed25329
[skip travis] explicitly include return type for all documented functions
