+++ beamshapes/point_source_on_a_sphere.py
@@ -46,8 +46,7 @@ docs_for_dtheta = '''On-axis level for point on a sphere.
                  
                  Returns 
                  -------
-                 _ : np.float
-                    Off-axis level at :math:`\\theta` for point-source on a sphere.
+                 _ : np.float, Off-axis level at :math:`\\theta` for point-source on a sphere.
                  '''
 d_theta_func.__doc__ = docs_for_dtheta
 
@@ -58,7 +57,8 @@ d_zero = (1/(kR**2))*onaxissum_term
 
 d_zero_func = lambdify([k,R,NN],d_zero)
 
-docs_for_dzero = '''On-axis level for point on a sphere. 
+docs_for_dzero = '''
+                 On-axis level for point on a sphere. 
                  Lambdified version of :code:`d_zero` SymPy expression.
                  
                  Parameters
@@ -72,8 +72,8 @@ docs_for_dzero = '''On-axis level for point on a sphere.
                  
                  Returns 
                  -------
-                 _ : np.float
-                    On-axis level for point-source on a sphere.
+                 _ : np.float,  On-axis level for point-source on a sphere.
+
                  '''
 
 d_zero_func.__doc__ = docs_for_dzero
