### 2021-12-07 thejasvibr/bat_beamshapes#22 @thejasvibr
Have addressed this issue. All documented functions now have the returned object and the types explicitly stated as of [ed2532](https://github.com/thejasvibr/bat_beamshapes/commit/ed25329abb46e480ed261593813231995f7a01ab).

Hope this sorts the deficiency in [the docs](https://beamshapes.readthedocs.io/en/latest/source_models.html)?
