### 2021-12-08 5877906
[skip travis] include a few more details in docs and MS
