+++ paper.md
@@ -41,7 +41,7 @@ Directivity functions can be used in a two-fold manner 1) to deliberately engine
 
 # Statement of need
 
-A host of published directivity functions exist in the literature, but it is my experience that their computational implementations remain as inhouse scripts that are often in proprietary language platforms. To my knowledge there are no computational implementations of multi-parameter sound source directivities. For instance, the ```levitate``` package [@levitate] implements directivities of 'simple' sound sources that can be defined with one parameter (e.g. piston radius in piston in a baffle, or effective radius for a circular ring). The directivities of simple source models can be rapidly implemented and computed as analytical solutions are often available. In contrast to simple source models, the directivities of other models (as in this package) involve more parameters (e.g. piston and sphere radius for piston in a sphere) and do not have analytical solutions. The directivity calculations require numerical routines and run-time optimisation. The advantage of more involved source models is however their ability to capture aspects of experimental sound sources. In this paper I present ```beamshapes```, a Python package that currently implements directivity patterns for two 'simple' and two 'involved' sound source models. As of this publication, ```beamshapes``` version 0.2.0 implements four sound sources 1) piston in an infinite baffle, 2) point source on a sphere 3) oscillating cap of a sphere and 4) piston in a sphere. 
+A host of published directivity functions exist in the literature, but it is my experience that their computational implementations remain as inhouse scripts that are often in proprietary language platforms. To my knowledge there are no computational implementations of multi-parameter sound source directivities. For instance, the ```levitate``` package [@levitate] implements directivities of 'simple' sound sources that can be defined with one parameter (e.g. piston in a baffle, circular ring). Since analytical solutions are available, the directivities of simple source models can be rapidly implemented and computed. In contrast to simple source models, the directivities of other models (as in this package) involve more parameters (e.g. piston in a sphere, oscillating cap of a sphere) and do not have analytical solutions. Their directivity calculations require numerical routines and run-time optimisation. The advantage of more involved source models is however their ability to capture aspects of experimental sound sources. In this paper I present ```beamshapes```, a Python package that currently implements directivity patterns for two 'simple' and two 'involved' sound source models. As of this publication, ```beamshapes``` version 0.2.0 implements four sound sources 1) piston in an infinite baffle, 2) point source on a sphere 3) oscillating cap of a sphere and 4) piston in a sphere. 
 
 Computational implementations of directivity functions often require long run-times due to the intensive numerical routines and arbitrary precision math involved. Long run-times hinder scientific projects in reducing the number of models and parameter space that can be explored. ```beamshapes``` boasts parallelised code to generate significant speed-ups in run-times. 
 
