### 2021-12-09 thejasvibr/bat_beamshapes#22 @hagenw
We still have the returned tuple with `None` as first entry

![image](https://user-images.githubusercontent.com/173624/145387336-509f7cab-8983-44e4-a34e-fdf67c21f3db.png)

How should I interpret this as a user? Is the first return value meaningless and I should call the function with:

```python
_, beamshapes = beamshapes.piston_in_infinite_baffle.piston_in_infinite_baffle_directivity(angles, params)
```
