### 2021-12-09 4ef42ae
[skip travis] alter docs to give correct return object for piston infinite baffle directivity
