### 2021-12-09 thejasvibr/bat_beamshapes#22 @thejasvibr
Yes, this is how it is intended (return variable name corrected to ```dtheta_by_dzero``` in the [latest commit](https://github.com/thejasvibr/bat_beamshapes/commit/4ef42aef0e0848275abd34c56e72e840378d53e6)).

The first returned object of *<source_model>_directivity* function is intended to be a ```None``` OR a matrix calculated for source models whose directivities take a long time to run. If the directivity needs to be calculated again for a new set of angles, the returned matrix can be used.

The first returned object is ```None``` for fast running directivities (piston in infinite baffle, cap of a sphere, point source on a sphere) - while it is[ a matrix for the piston in a sphere](https://beamshapes.readthedocs.io/en/latest/source_models.html#beamshapes.piston_in_sphere.piston_in_sphere_directivity). In general, as more time consuming source models are added, I expect the user will find the calculated matrix a major time-saver.

For example - if one were to run a piston in a sphere - it'd typically take 2-5 mins to get directivities for one set of angles:

```
import beamshape
matrix_coefs, set1_directivities = beamshape.piston_in_sphere_directivity(set1_angles, params) # this takes 2-5 min for e.g.

# now say we want to calculate the directivity for another set of angles - the calculation happens very quickly
#because we include the pre-calculated matrix
_, set2_directivities = beamshape.piston_in_sphere_directivity(set2_angles, params, A_n=matrix_coefs)
```
