### 2021-12-14 thejasvibr/bat_beamshapes#7 @nils-werner
Yes, the docs are much better now, thanks! (you could also link to them in your repo settings)
