### 2021-12-14 thejasvibr/bat_beamshapes#8 @nils-werner
Yes, this is OK. As I said, it wouldn't be a reason for rejection.
