### 2021-12-15 thejasvibr/bat_beamshapes#19 @hagenw
Cool, thank you, this is now much clearer and provides a good overview what is available in your package.
