### 2021-12-15 #3740 @faroit
> For me it is shown as ticked:
>
>
>
> ![image](https://user-images.githubusercontent.com/173624/146152819-1ff07e53-6566-466f-857b-d67366c4792d.png)
>
>

Great. Thanks maybe my browser hasn't updated its cache.
