### 2021-12-15 0a2cea7
[skip travis] add pyroomacoustics DOI to references
