+++ paper.bib
@@ -44,6 +44,7 @@
   pages={351--355},
   year={2018},
   organization={IEEE}
+  doi={10.1109/ICASSP.2018.8461310}
 }
 
 @article{mogensen1979sound,
