### 2021-12-15 9c2919b
[skip travis] correct pyroomacoustics bibtex entry
