+++ paper.bib
@@ -43,7 +43,7 @@
   booktitle={2018 IEEE International Conference on Acoustics, Speech and Signal Processing (ICASSP)},
   pages={351--355},
   year={2018},
-  organization={IEEE}
+  organization={IEEE},
   doi={10.1109/ICASSP.2018.8461310}
 }
 
