### 2022-01-04 #3740 @faroit
This submission looks good now and we can finally move forward 👏 I just did check the paper again for grammar and spellings and created a pull request that you can merge in https://github.com/thejasvibr/bat_beamshapes/pull/24 if you like.

@thejasvibr when the paper is ready, please make a new (tagged) release of the main repo that includes all of the changes that have resulted from the review. Please report the version number here.

Then, please create an archive of the software in Zenodo/figshare/other service and update this thread with the DOI of the archive. For the archive version, please make sure that:

- The authors of the archive are the same as the JOSS paper authors
