### 2022-01-05 #3740 @thejasvibr
Great!
* The paper + repo is ready with a tagged release (v0.2.2) on the Github repo.

* The same release archive has been uploaded as .zip and .tar.gz on Zenodo - with the following DOI: 10.5281/zenodo.5819771 ([link to Zenodo deposition](https://doi.org/10.5281/zenodo.5819771))
