### 2022-01-05 #3740 @thejasvibr
* The PDF proof looks good 😎
* The XML file however didn't seem to parse correctly when uploaded at the Crossref testing page ([link](https://data.crossref.org/reports/parser.html)). The error output is below:

![image](https://user-images.githubusercontent.com/12443569/148217897-c7cc5196-c2c9-496d-9395-c5e04506f13e.png)

@faroit, I have no knowledge of how Crossref XMLs should look like, not sure if this is a serious error or not ...
