### 2022-01-05 #3740 @kthyng
@thejasvibr I'll do some final checks before publishing. Actually the XML does look weird.

@openjournals/dev The XML doesn't have very many lines  — is there an issue?
