### 2022-01-05 #3740 @kthyng
@thejasvibr -

* version and archive look good ✅
* paper:
- You don't mention the package name in the Summary but you should.
- Under "Software..." section, please list the reference for each package immediately after each package, and format package name with ``.
- The "Package repository" section is a bit redundant with the left hand side of the paper listing the repo location.
- References: Shouldn't Brumm 2002 have the species name be italicized?
- Gerhardt 1975 should have "north american" capitalized. You can preserve capitalization with {} around characters.
- please check other reference capitalization
