### 2022-01-06 #3740 @thejasvibr
Thanks @kthyng - will do the reformatting. How to proceed now, should I make another tagged release, or just make the manuscript changes into a normal commit?
