### 2022-01-06 1176c01
add changes to capitalisation and italics in BibTex
