+++ paper.bib
@@ -207,7 +207,7 @@
 }
 
 @article{gerhardt1975sound,
-  title={Sound pressure levels and radiation patterns of the vocalizations of some North American frogs and toads},
+  title={Sound pressure levels and radiation patterns of the vocalizations of some {N}orth {A}merican frogs and toads},
   author={Gerhardt, H Carl},
   journal={Journal of Comparative Physiology},
   volume={102},
@@ -278,7 +278,7 @@
 
 
 @article{brumm2002sound,
-  title={Sound radiation patterns in nightingale (Luscinia megarhynchos) songs},
+  title={Sound radiation patterns in nightingale (\textit{{L}uscinia megarhynchos}) songs},
   author={Brumm, Henrik},
   journal={Journal f{\"u}r Ornithologie},
   volume={143},
