### 2022-01-06 4982a12
1) include package name in summary 2) include further italics and caps in references
