+++ paper.bib
@@ -196,7 +196,7 @@
 }
 
 @article{rodriguez2020sound,
-  title={Sound radiation pattern of the advertisement call of the highly territorial poison frog Allobates femoralis},
+  title={Sound radiation pattern of the advertisement call of the highly territorial poison frog \textit{{A}llobates femoralis},
   author={Rodr{\'\i}guez, Camilo and H{\"o}dl, Walter},
   journal={Behavioural processes},
   volume={170},
@@ -219,7 +219,7 @@
 }
 
 @article{patricelli2008acoustic,
-  title={Acoustic directionality of red-winged blackbird (Agelaius phoeniceus) song relates to amplitude and singing behaviours},
+  title={Acoustic directionality of red-winged blackbird (\textit{{A}gelaius phoeniceus}) song relates to amplitude and singing behaviours},
   author={Patricelli, Gail L and Dantzker, Marc S and Bradbury, Jack W},
   journal={Animal Behaviour},
   volume={76},
