+++ paper.bib
@@ -196,7 +196,7 @@
 }
 
 @article{rodriguez2020sound,
-  title={Sound radiation pattern of the advertisement call of the highly territorial poison frog \textit{{A}llobates femoralis},
+  title={Sound radiation pattern of the advertisement call of the highly territorial poison frog \textit{{A}llobates femoralis}},
   author={Rodr{\'\i}guez, Camilo and H{\"o}dl, Walter},
   journal={Behavioural processes},
   volume={170},
