### 2022-01-06 8e70981
include further changes in referencing + italics
