+++ paper.bib
@@ -265,7 +265,7 @@
 }
 
 @article{patricelli2007differences,
-  title={Differences in acoustic directionality among vocalizations of the male red-winged blackbird (Agelaius pheoniceus) are related to function in communication},
+  title={Differences in acoustic directionality among vocalizations of the male red-winged blackbird (\textit{{A}gelaius pheoniceus)} are related to function in communication},
   author={Patricelli, Gail L and Dantzker, Marc S and Bradbury, Jack W},
   journal={Behavioral Ecology and Sociobiology},
   volume={61},
