### 2022-01-06 #3740 @kthyng
@thejasvibr Your changes look good! You can keep these changes in the separate branch if you want, or merge, either is fine. You don't need to update the Zenodo archive or version number for these small manuscript changes unless you want to. Let me know when you're ready to proceed and we'll wrap it up!
