### 2022-01-07 #3740 @thejasvibr
Proof looks good to proceed @kthyng, thanks for the clarifications! Will keep the paper on the ```ms``` branch for now.
