### 2022-01-08 #3740 @danielskatz
@openjournals/dev - can you see what the problem is here - recommend-accept from the branch doesn't work, but generate pdf from the branch does
