### 2022-01-09 #3740 @thejasvibr
Thanks for checking @danielskatz @openjournals/dev . Have deleted the ```ms``` branch, and now the package + latest paper version are in the same branch (```dev```). Hopefully this will lead to better luck?
