### 2022-01-09 #3740 @danielskatz
👋 @thejasvibr - a few more small changes are in https://github.com/thejasvibr/bat_beamshapes/pull/25 - if you can merge this, I can proceed to publication
