### 2022-01-10 #3740 @danielskatz
@thejasvibr - I'm waiting for the DOI to resolve before I will close this issue and declare the paper publisher.  While this usually happens quickly, sometimes it can take a little bit.

@arfon - any thoughts?
