### 2022-01-10 #3740 @arfon
> @arfon - any thoughts?

Looks like this line was breaking the XML generation https://github.com/thejasvibr/bat_beamshapes/blob/dev/paper.bib#L186. I've fixed this manually and redeposited the XML with Crossref.
